refacto: remove flag support
Now that Zenoh-Flow supports composite operators, the need for Flags is
drastically reduced. This commit drops their support all together.

The main motivation behind this design is to promote "better" design: instead of
depending on Flags to change a data flow, creating several distinct flows
that leverage composition is preferred.